+++ core/src/main/java/org/jdbi/v3/core/Handle.java
@@ -410,7 +410,9 @@ public class Handle implements Closeable, Configurable<Handle> {
      * @throws X any exception thrown by the callback
      */
     public <R, X extends Exception> R inTransaction(HandleCallback<R, X> callback) throws X {
-        return transactions.inTransaction(this, callback);
+        return isInTransaction()
+            ? callback.withHandle(this)
+            : transactions.inTransaction(this, callback);
     }
 
     /**
@@ -422,7 +424,7 @@ public class Handle implements Closeable, Configurable<Handle> {
      * @throws X any exception thrown by the callback
      */
     public <X extends Exception> void useTransaction(final HandleConsumer<X> callback) throws X {
-        transactions.inTransaction(this, handle -> {
+        inTransaction(handle -> {
             callback.useHandle(handle);
             return null;
         });
@@ -446,6 +448,16 @@ public class Handle implements Closeable, Configurable<Handle> {
      * @throws X any exception thrown by the callback
      */
     public <R, X extends Exception> R inTransaction(TransactionIsolationLevel level, HandleCallback<R, X> callback) throws X {
+        if (isInTransaction()) {
+            TransactionIsolationLevel currentLevel = getTransactionIsolationLevel();
+            if (currentLevel != level && level != TransactionIsolationLevel.UNKNOWN) {
+                throw new TransactionException(
+                    "Tried to execute nested transaction with isolation level " + level + ", "
+                    + "but already running in a transaction with isolation level " + currentLevel + ".");
+            }
+            return callback.withHandle(this);
+        }
+
         try (TransactionResetter tr = new TransactionResetter(getTransactionIsolationLevel())) {
             setTransactionIsolation(level);
             return transactions.inTransaction(this, level, callback);