+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/transaction/internal/TransactionDecorator.java
@@ -33,18 +33,9 @@ public class TransactionDecorator implements HandlerDecorator {
         return (target, args, handle) -> {
             Handle h = handle.getHandle();
 
-            if (h.isInTransaction()) {
-                // Already in transaction. The outermost @Transaction method determines the transaction isolation level.
-                TransactionIsolationLevel currentLevel = h.getTransactionIsolationLevel();
-                if (currentLevel != isolation && isolation != TransactionIsolationLevel.UNKNOWN) {
-                    throw new TransactionException("Tried to execute nested @Transaction(" + isolation + "), "
-                            + "but already running in a transaction with isolation level " + currentLevel + ".");
-                }
-                if (h.isReadOnly() && !readOnly) {
-                    throw new TransactionException("Tried to execute a nested @Transaction(readOnly=false) "
-                            + "inside a readOnly transaction");
-                }
-                return base.invoke(target, args, handle);
+            if (h.isInTransaction() && h.isReadOnly() && !readOnly) {
+                throw new TransactionException("Tried to execute a nested @Transaction(readOnly=false) "
+                        + "inside a readOnly transaction");
             }
 
             HandleCallback<Object, Exception> callback = th -> base.invoke(target, args, handle);