+++ RELEASE_NOTES.md
@@ -21,6 +21,14 @@
   - Rework caching to not use extra threads (#1453)
   - Any valid Java identifier is now supported as a named parameter (e.g. `:제목`) or defined attribute
     (e.g. `<제목>`).
+  - Nested `inTransaction` and `useTransaction` calls are now allowed in core, provided the inner
+    transaction specifies the same transaction isolation level, or does not specify isolation. This
+    brings core transaction behavior in line with the existing behavior for nested `@Transaction` SQL
+    object methods.
+  - Nested calls on a thread to `Jdbi` methods `useHandle`, `withHandle`, `useTransaction`,
+    `withTransaction`, `useExtension`, `withExtension`, or to any method of an on-demand extension will
+    now execute against the handle from the outermost call, rather than each invocation getting a separate
+    handle.
 - Minor source incompatibility
   - JdbiPlugin methods now `throws SQLException`